add context for logo change to immediately update logo
