+++ frontend/src/hooks/useLogo.js
@@ -1,22 +1,7 @@
-import { useEffect, useState } from "react";
-import System from "../models/system";
-import AnythingLLM from "../media/logo/anything-llm.png";
+import { useContext } from "react";
+import { LogoContext } from "../LogoContext";
 
 export default function useLogo() {
-  const [logo, setLogo] = useState("");
-
-  useEffect(() => {
-    async function fetchInstanceLogo() {
-      try {
-        const logoURL = await System.fetchLogo();
-        logoURL ? setLogo(logoURL) : setLogo(AnythingLLM);
-      } catch (err) {
-        setLogo(AnythingLLM);
-        console.error("Failed to fetch logo:", err);
-      }
-    }
-    fetchInstanceLogo();
-  }, []);
-
-  return { logo };
+  const { logo, setLogo } = useContext(LogoContext);
+  return { logo, setLogo };
 }
